+++ shared/contactdatabase/src/main/java/com/dmitrystonie/contactbook/contactdatabase/data/entity/PictureEntity.kt
@@ -5,22 +5,23 @@ import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
 
-@Entity(tableName = PictureEntity.TABLE_NAME,     foreignKeys = [ForeignKey(
-    entity = ContactEntity::class,
-    parentColumns = [ContactEntity.ID_COLUMN_NAME],
-    childColumns = [CoordinatesEntity.CONTACT_ID_NAME],
-    onDelete = ForeignKey.CASCADE,
-)],)
+@Entity(
+    tableName = PictureEntity.TABLE_NAME,
+    foreignKeys = [ForeignKey(
+        entity = ContactEntity::class,
+        parentColumns = [ContactEntity.ID_COLUMN_NAME],
+        childColumns = [CoordinatesEntity.CONTACT_ID_NAME],
+        onDelete = ForeignKey.CASCADE,
+    )],
+)
 data class PictureEntity(
-    @PrimaryKey(autoGenerate = true)
-    val id: Int,
-    @ColumnInfo(name = CONTACT_ID_NAME, index = true)
-    val contactId: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long,
+    @ColumnInfo(name = CONTACT_ID_NAME, index = true) val contactId: Long,
     val largeUrl: String,
     val mediumUrl: String,
     val thumbnailUrl: String,
-){
-    companion object{
+) {
+    companion object {
         const val TABLE_NAME = "pictures"
         const val CONTACT_ID_NAME = "contact_id"
